refactor: extract useAuth and AuthContext to lib/auth-context
